+++ src/main/java/dev/muon/otherworldorigins/network/RespecAptitudesMessage.java
@@ -4,13 +4,19 @@ import com.seniors.justlevelingfork.common.capability.AptitudeCapability;
 import com.seniors.justlevelingfork.network.packet.client.SyncAptitudeCapabilityCP;
 import com.seniors.justlevelingfork.network.packet.common.AptitudeLevelUpSP;
 import com.seniors.justlevelingfork.registry.RegistryAptitudes;
+import com.seniors.justlevelingfork.registry.RegistryPassives;
 import com.seniors.justlevelingfork.registry.RegistrySkills;
+import com.seniors.justlevelingfork.registry.RegistryTitles;
 import com.seniors.justlevelingfork.registry.aptitude.Aptitude;
+import com.seniors.justlevelingfork.registry.passive.Passive;
 import com.seniors.justlevelingfork.registry.skills.Skill;
+import com.seniors.justlevelingfork.registry.title.Title;
 import dev.muon.otherworldorigins.OtherworldOrigins;
 import dev.muon.otherworldorigins.power.InnateAptitudeBonusPower;
+import net.minecraft.core.Registry;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.entity.player.Player;
 import net.minecraftforge.network.NetworkEvent;
 
 import java.util.function.Supplier;
@@ -50,12 +56,16 @@ public class RespecAptitudesMessage {
 
                     resetAptitudes(capability, player);
                     resetSkills(capability);
-                    ResetFeatsMessage.handle(new ResetFeatsMessage(), ctx); // Reset feats
+                    resetPassives(capability);
+                    resetTitles(capability);
 
+                    ResetFeatsMessage.handle(new ResetFeatsMessage(), ctx);
+                    RegistryTitles.syncTitles(player);
                     SyncAptitudeCapabilityCP.send(player);
+
                     OtherworldOrigins.LOGGER.debug("Respec completed successfully");
                 } else {
-                    OtherworldOrigins.LOGGER.warn("Player doesn't have enough levels");
+                    OtherworldOrigins.LOGGER.warn("Player {} attempted to respec without sufficient XP. Obliterate them", player);
                 }
             }
         });
@@ -87,18 +97,25 @@ public class RespecAptitudesMessage {
         }
     }
 
-
-    private static void resetAptitudes(AptitudeCapability capability) {
-        for (Aptitude aptitude : RegistryAptitudes.APTITUDES_REGISTRY.get().getValues()) {
-            capability.setAptitudeLevel(aptitude, 1);
-        }
-    }
-
     private static void resetSkills(AptitudeCapability capability) {
         for (Skill skill : RegistrySkills.SKILLS_REGISTRY.get().getValues()) {
             capability.setToggleSkill(skill, false);
         }
     }
+
+    private static void resetPassives(AptitudeCapability capability) {
+        for (Passive passive : RegistryPassives.PASSIVES_REGISTRY.get().getValues()) {
+            capability.subPassiveLevel(passive, 10);
+        }
+    }
+
+    private static void resetTitles(AptitudeCapability capability) {
+        for (Title title : RegistryTitles.TITLES_REGISTRY.get().getValues()) {
+            capability.setUnlockTitle(title, false);
+        }
+        capability.setPlayerTitle(RegistryTitles.getTitle("titleless"));
+    }
+
 
     public static void send() {
         OtherworldOrigins.CHANNEL.sendToServer(new RespecAptitudesMessage());